feat(chan): Use our own io->chan implementation
The lsp4clj implementation is for the LSP protocol, which has the
concept of "header" lines. For reasons that I do not understand, MCP
does not adopt this. This is why my server was not working. This file
has an initial implementation of a reader/writer that speaks MCP. It's
still broken, but we are on our way to figuring out why.