feat: Add connected-clients to the global state
This is a necessary first step towards implementing dynamic capability
negotiation.

We also add a note about why our empty-context-map is so shitty, why
there is so much internal mutable state in it. I want to optimize this
in the future, once the nature of the possible changes becomes better
understood. Leaving it as-is for the moment.