feat(negotiation): Add protocol version negotiation for client/server
Other than returning the correct version, this commit does not
introduce anything else right now. Even so, this is important because
we are implementing correctness of the spec.

(Yes, I see the irony of the statement above, given that I am not
currently supporting a large part of both published versions of the
spec. The hope is to complete the whole damn spec before 1st June,
releasing a 1.0.0 version)